Hotfix: Update to work with newer versions of ruby
This updates the gem to work with newer versions of ruby. This
didn't require any real changes to the code. The only changes
needed were adding the gem dir to the load path for the
bench/examples scripts and adding the `choice` backfill for ruby
1.8.7. We aren't currently using `choice` in the test suite but
this updates to match how GGem generates the `test/helper`.